myit/requirements.txt
# myit MVP Dependencies
# Backend
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
aiosqlite==0.20.0
alembic==1.14.0

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.18

# AI / Embedding
google-generativeai==0.8.4
chromadb==0.5.23
sentence-transformers==3.3.1

# Data Collection
httpx==0.28.1
feedparser==6.0.11
beautifulsoup4==4.12.3
lxml==5.3.0

# Utils
pydantic-settings==2.7.0
python-dotenv==1.0.1
pandas==2.2.3

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
httpx  # also used for testing
